requirements.txt
absl-py==1.4.0
attrs==23.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.24.3
open-spiel==1.3
pydantic==1.10.9
Pygments==2.15.1
rich==13.4.2
scipy==1.10.1
typing_extensions==4.6.3
